fix compile time warnings (#352)

+++ sdk/src/metrics/meter.cc
@@ -547,7 +547,7 @@ void Meter::RecordShortBatch(const trace::KeyValueIterable &labels,
                              nostd::span<metrics_api::SynchronousInstrument<short> *> instruments,
                              nostd::span<const short> values) noexcept
 {
-  for (int i = 0; i < instruments.size(); ++i)
+  for (size_t i = 0; i < instruments.size(); ++i)
   {
     instruments[i]->update(values[i], labels);
   }
@@ -557,7 +557,7 @@ void Meter::RecordIntBatch(const trace::KeyValueIterable &labels,
                            nostd::span<metrics_api::SynchronousInstrument<int> *> instruments,
                            nostd::span<const int> values) noexcept
 {
-  for (int i = 0; i < instruments.size(); ++i)
+  for (size_t i = 0; i < instruments.size(); ++i)
   {
     instruments[i]->update(values[i], labels);
   }
@@ -567,7 +567,7 @@ void Meter::RecordFloatBatch(const trace::KeyValueIterable &labels,
                              nostd::span<metrics_api::SynchronousInstrument<float> *> instruments,
                              nostd::span<const float> values) noexcept
 {
-  for (int i = 0; i < instruments.size(); ++i)
+  for (size_t i = 0; i < instruments.size(); ++i)
   {
     instruments[i]->update(values[i], labels);
   }
@@ -577,7 +577,7 @@ void Meter::RecordDoubleBatch(const trace::KeyValueIterable &labels,
                               nostd::span<metrics_api::SynchronousInstrument<double> *> instruments,
                               nostd::span<const double> values) noexcept
 {
-  for (int i = 0; i < instruments.size(); ++i)
+  for (size_t i = 0; i < instruments.size(); ++i)
   {
     instruments[i]->update(values[i], labels);
   }
@@ -738,7 +738,7 @@ bool Meter::IsValidName(nostd::string_view name)
     return false;
   else
   {
-    for (int i = 0; i < name.size(); ++i)
+    for (size_t i = 0; i < name.size(); ++i)
     {
       if (!isalnum(name[i]) && name[i] != '_' && name[i] != '.' && name[i] != '-')
         return false;
